refactor(Test) : 테스트 코드 클래스 별 분리
